requirements.txt
numpy==1.26.0
opencv-python==4.8.0.76
packaging==23.2
Pillow==10.0.1
pygame==2.5.2
pyserial==3.5
pytesseract==0.3.10
